Refactor: reorganización de Navbar
Se actualiza la barra de navigation para su uso en el proyecto.
